Many minor changes
- Windows bat improved
- Type cleanups
- Little Endian color fix
NOTE: Crashes on my Mac
NOTE: Stereogram errors on Windows

+++ examinationroom/src/core/texture.h
@@ -16,6 +16,7 @@
 #include <memory>
 
 #include <QImage>
+#include <qgl.h>
 
 namespace Examination
 {
@@ -70,7 +71,7 @@ private:
 	QImage image_;
 	QImage original_;
 	
-	unsigned long imageGLID_;
+	GLuint imageGLID_;
 };
 
 }
